lots of stuff whoops

+++ api/models/feed_like.php
@@ -35,7 +35,7 @@ class feedLikeModel
     public function delete()
     {
         if (!isset($this->feedid)){$this->app->errorJSON($this->app, "No ID was selected to delete");}
-        $this->app->db->delete("feed_likes", ["feedid" => $this->feedid]);
+        $this->app->db->delete("feed_likes", ["AND" => ["feedid" => $this->feedid, "userid" => $this->userid]]);
  
     }
